lint(app): fix all lint errors
Issue #8
