+++ package.json
@@ -6,7 +6,7 @@
   "main": "index.js",
   "scripts": {
     "sequelize": "node_modules/.bin/sequelize",
-    "lint": "eslint --ext .js,.jsx app server",
+    "lint": "eslint --ext .js,.jsx src",
     "lint:fix": "eslint --ext .js,.jsx app server --fix",
     "clean": "rimraf public && rimraf compiled",
     "mobile": "node node_modules/react-native/local-cli/cli.js start",
@@ -22,7 +22,6 @@
   "devDependencies": {
     "babel-jest": "18.0.0",
     "babel-preset-react-native": "1.9.1",
-    "chai": "^3.5.0",
     "enzyme": "^2.4.1",
     "eslint": "^3.5.0",
     "eslint-config-airbnb": "^10.0.0",
@@ -69,6 +68,7 @@
     "bcrypt-nodejs": "0.0.3",
     "bluebird": "^3.4.6",
     "body-parser": "^1.15.0",
+    "chai": "^3.5.0",
     "classnames": "^2.2.3",
     "compression": "^1.6.2",
     "connect-mongo": "^1.3.2",
