+++ src/mobile/config/StyleConst.js
