+++ src/server/db/mongo/controllers/users.controller.js
@@ -1,9 +1,22 @@
 import R from 'ramda';
-import User from '../models/user.model';
-import {then, catchErr} from '../../../utils/invokers';
-import {promisify} from '../../../utils/monads';
 import jwt from 'jsonwebtoken';
+import User from '../models/user.model';
+import {then, catchErr} from '../../../../utils/invokers';
+import {promisify} from '../../../../utils/monads';
 import {SECRET_KEY} from '../../../../../config/env';
+
+// Private part
+
+const generateToken = user => ({
+  user,
+  token: jwt.sign(user._id.toString(), SECRET_KEY)
+});
+
+const getUserIdByToken = promisify(token => jwt.verify(token, SECRET_KEY));
+
+const getToken = req => req.get('token');
+
+// Public part
 
 export const extractUserMiddleware = (req, res, next) => R.pipe(
   getToken,
@@ -11,11 +24,11 @@ export const extractUserMiddleware = (req, res, next) => R.pipe(
   then(userId => User.findById(userId)),
   then(
     user => {
-      req.user = user;
+      req.user = user; // eslint-disable-line no-param-reassign
       next();
     }
   ),
-  catchErr(err => next())
+  catchErr(() => next())
 )(req);
 
 export const signUp = (_, args) => R.pipe(
@@ -31,14 +44,3 @@ export const signIn = (_, args) => R.pipe(
 )(args);
 
 export const me = (_, __, context) => context.user;
-
-// Private part
-
-const generateToken = user => ({
-  user,
-  token: jwt.sign(user._id.toString(), SECRET_KEY)
-});
-
-const getUserIdByToken = promisify(token => jwt.verify(token, SECRET_KEY));
-
-const getToken = req => req.get('token');
