+++ src/server/index.js
@@ -14,13 +14,12 @@ const app = express();
  */
 connect();
 
-
-
 if (isDebug) {
   // enable webpack hot module replacement
   const webpackDevMiddleware = require('webpack-dev-middleware');
   const webpackHotMiddleware = require('webpack-hot-middleware');
   const webpackConfig = require('../../webpack/webpack.config');
+
   const devBrowserConfig = webpackConfig('browser');
   const compiler = webpack(devBrowserConfig);
   app.use(webpackDevMiddleware(compiler, { noInfo: true, publicPath: devBrowserConfig.output.publicPath }));
