+++ src/server/init/routes.js
@@ -1,6 +1,3 @@
-/**
- * Routes for express app
- */
 import graphqlHTTP from 'express-graphql';
 import schema from '../schema/schema';
 import {extractUserMiddleware} from '../db/mongo/controllers/users.controller';
@@ -10,7 +7,7 @@ export default (app) => {
     '/graphql',
     extractUserMiddleware,
     graphqlHTTP({
-      schema: schema,
+      schema,
       graphiql: true,
     })
   );
