--- src/server/utils/constructors.js
@@ -1,3 +1,0 @@
-import R from 'ramda';
-
-export const PromiseConstructor = R.construct(Promise);
