+++ src/utils/constructors.js
@@ -1,0 +1,3 @@
+import R from 'ramda';
+
+export const promiseConstructor = R.construct(Promise); // eslint-disable-line import/prefer-default-export
